#163: Moved from XNAssets to AssetManagementBase

+++ src/Myra/Utility/CrossEngineStuff.cs
@@ -15,7 +15,7 @@ namespace Myra.Utility
 		public static Color CreateColor(int r, int g, int b, int a = 255)
 		{
 #if MONOGAME || FNA || STRIDE
-			return new Color(r, g, b, a);
+			return new Color((byte)r, (byte)g, (byte)b, (byte)a);
 #else
 			return Color.FromArgb(a, r, g, b);
 #endif
